mangaba_service/requirements.txt
mangaba_ai
fastapi
uvicorn
python-dotenv
